Copy SecureEdit from KeePass 2.38
KeePass 2.39 is dropping the secure edit control.

See: https://sourceforge.net/p/keepass/discussion/329220/thread/4236e391/#a6fa

This copies the SecureEdit class from KeePass 2.38 and puts in in the KeePassRPC namespace. The uses of this control are changed to qualify the namespace so that this compiles on versions of KeePass with or without KeePass.UI.SecureEdit.