WIP improvements to HTTP Auth filtering Realms have been optional since 2014 so some websites may not include them. These websites will not function correctly with the current version of Kee and this branch. Specifically, matched entries will not be demoted to a lower quality match even if the entry has a missing or incorrect realm defined.
Since Kee does not do anything with the match quality information for HTTP Auth at the moment, this point, and this entire branch is moot.

These LST values are currently sent from Kee (other clients may be different though)
LSTnoForms if realm is set and not empty ===== 2
LSTnoRealms only if realm is not set or is empty ===== 1

It's likely that a change to the findLogins API will be needed to implement this accurately. For example, a new parameter which indicates the source of the request rather than the current "type of search to exclude" inherited from Firefox 10 years ago.